fix: Update CMakeLists.txt to include 'ICU::uc' component for dcc-datetime-plugin
This commit updates the ICU components in the CMakeLists.txt file. Previously,
only the 'i18n' component was being found using find_package for ICU. This
update includes the 'uc' component as well to ensure that all necessary ICU
components are found and linked correctly during the build process.

By adding 'ICU::uc' to the list of components, we ensure that the DateTime
plugin can make use of the required ICU functionalities without any issues.