fix: ScorllBar still hide when hovered
Contorl will eat mouse presse event, and it causes window can't
be dragged, That's why the Control is disabled hoverEnabled in
DccWindow.
TitleBar maybe assigned to header of Application, but it causes
SideBar is not conforming to the design ui.
hoverEnabled has an inheritance relationship.
We will release the hoverEnabled at the root of the content.

pms: BUG-311835